[Composer, Sonnet reviewed plan] Stop using a cloned Range held in React state
